requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
python-dotenv==1.0.0
PyJWT==2.8.0

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# AI & Machine Learning
google-generativeai==0.3.2
ibm-watson-machine-learning==1.0.335
transformers==4.38.0
torch==2.1.0
ollama==0.5.3
huggingface-hub==0.19.4
tokenizers==0.15.0

# PDF Generation & Document Processing
reportlab==4.0.7
PyPDF2==3.0.1

# Image Processing
Pillow==10.1.0

# Server (Optional - for production)
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.7.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Logging & Monitoring
structlog==23.2.0

# Caching
redis==5.0.1

# Email Support (Optional)
Flask-Mail==0.9.1

# File Upload
Flask-Uploads==0.2.1

# Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# Background Tasks (Optional - for advanced features)
celery==5.3.4
redis==5.0.1

# Health Checks
healthcheck==1.3.3

# Rate Limiting
Flask-Limiter==3.5.0

# CORS
Flask-CORS==4.0.0

# Environment Management
python-decouple==3.8

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# JSON Processing
jsonschema==4.20.0

# Timezone Support
pytz==2023.3

# Compression
Flask-Compress==1.14

# Session Management
Flask-Session==0.5.0

# Error Handling (Optional - for production monitoring)
sentry-sdk[flask]==1.38.0

# Performance Monitoring (Optional)
flask-profiler==1.8

# API Documentation
flasgger==0.9.7.1

# Database Drivers (Optional)
psycopg2-binary==2.9.9  # PostgreSQL
pymysql==1.1.0  # MySQL

# Development Tools
ipython==8.17.2
jupyter==1.0.0
